autodetect keepalive (breaking change) (#44)
- change in the interface

+++ src/ChickenAPI.Packets/Interfaces/IPacket.cs
@@ -3,6 +3,6 @@
     public interface IPacket
     {
         string Header { get; set; }
-        int? KeepAliveId { get; set; }
+        ushort? KeepAliveId { get; set; }
     }
 }
